Module ic_artifact_manager::manager

The artifact manager implementation.

The artifact manager component manages all the artifact pools (Consensus pool, ingress pool, state sync pool, DKG pool, and certification pool).

It provides an interface to Gossip enabling it to interact with all the pools without knowing artifact-related details.

Structs

ArtifactManagerImpl
The artifact manager maintains a list of artifact clients, and is generic in the client type. It mostly just forwards function calls to each client depending on the artifact type.

ArtifactManagerMaker
The ArtifactManagerMaker is a helper to create an ArtifactManager after adding each client. It is separated from the ArtifactManager interface to ensure that all clients are added only once, and that the ArtifactManager can not be modified after creation.